requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
pydantic-settings==2.1.0
python-dotenv==1.0.0

cryptography==41.0.7

elasticsearch==8.11.0

sqlalchemy==2.0.23
psycopg2-binary==2.9.9
alembic==1.13.0

httpx==0.25.2
requests==2.31.0

python-json-logger==2.0.7

python-multipart==0.0.6
